requirements.txt
Python==3.11

Flask==3.1.2
flask-cors==6.0.1
pillow==11.3.0
numpy==1.26.4
tensorflow-cpu==2.16.2
keras==3.4.1

이 밑은 pip freeze > requirements.txt 로 얻은 것들임

absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.3.0
contourpy==1.3.3
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==24.3.25
fonttools==4.60.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.2
keras==3.4.1
keyring==23.5.0
kiwisolver==1.4.9
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==8.10.0
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.0
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pillow==11.3.0
protobuf==4.25.3
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.7.1
scikit-learn==1.7.2
scipy==1.16.2
SecretStorage==3.3.1
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.6.0
typing_extensions==4.12.2
urllib3==2.2.2
wadllib==1.3.6
Werkzeug==3.1.3
wrapt==1.16.0
zipp==1.0.0
